增加异步机制封装，增加windows环境支持。

+++ libuv_cpp11/Clinet.h
@@ -2,6 +2,7 @@
 #define CLIENT_H
 
 #include <iostream>
+#include <string>
 #include "uv/TcpClient.h"
 
 
@@ -36,7 +37,7 @@ public:
         }
     }
 
-    void newMessage(const char* buf,ssize_t size)
+    void newMessage(const char* buf,unsigned int size)
     {
         write(buf,size,nullptr);
     }
